requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--no-binary psycopg2

bleach==3.1.4             # via -r requirements.in
certifi==2019.11.28       # via requests, sentry-sdk
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.8         # via social-auth-core
defusedxml==0.6.0         # via python3-openid, social-auth-core
django-admin-autocomplete-filter==0.4  # via -r requirements.in
django-allauth==0.41.0    # via -r requirements.in
django-anymail==7.0.0     # via -r requirements.in
django-appconf==1.0.4     # via django-image-cropping
django-cleanup==4.0.0     # via -r requirements.in
django-cors-headers==3.2.1  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==2.2.9  # via -r requirements.in
django-filter==2.2.0      # via -r requirements.in
django-haystack==2.8.1    # via -r requirements.in
django-helusers==0.5.6    # via -r requirements.in
django-image-cropping==1.3.0  # via -r requirements.in
django-jinja==2.6.0       # via -r requirements.in
django-js-asset==1.2.2    # via django-mptt
django-leaflet==0.26.0    # via -r requirements.in
django-modeltranslation==0.14.4  # via -r requirements.in
django-mptt==0.11.0       # via -r requirements.in, django-munigeo, django-orghierarchy
django-munigeo==0.2.26     # via -r requirements.in
django-orghierarchy==0.1.24  # via -r requirements.in
django-parler-rest==2.1   # via django-munigeo
django-parler==2.0.1      # via django-munigeo, django-parler-rest
django-reversion==3.0.7   # via -r requirements.in
django==2.2.11            # via -r requirements.in, django-allauth, django-anymail, django-appconf, django-cors-headers, django-filter, django-haystack, django-helusers, django-jinja, django-leaflet, django-modeltranslation, django-mptt, django-munigeo, django-reversion, djangorestframework, djangorestframework-bulk, drf-oidc-auth, easy-thumbnails
djangorestframework-bulk==0.2.1  # via -r requirements.in
djangorestframework-gis==0.15  # via -r requirements.in
djangorestframework-jwt==1.11.0  # via -r requirements.in
djangorestframework==3.11.0  # via -r requirements.in, django-orghierarchy, django-parler-rest, djangorestframework-bulk, djangorestframework-gis, drf-oidc-auth
drf-oidc-auth==0.9        # via django-helusers
easy-thumbnails==2.7      # via -r requirements.in
ecdsa==0.15               # via python-jose
elasticsearch==7.6.0      # via -r requirements.in
future==0.18.2            # via pyjwkest
httmock==1.3.0            # via -r requirements.in
icalendar==4.0.5          # via -r requirements.in
idna==2.9                 # via requests
isodate==0.6.0            # via -r requirements.in, rdflib
jinja2==2.11.1            # via -r requirements.in, django-jinja
langdetect==1.0.8         # via -r requirements.in
lxml==4.5.0               # via -r requirements.in, python-docx
markdown==3.2.1           # via -r requirements.in
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
pillow==7.0.0             # via -r requirements.in, easy-thumbnails
psycopg2==2.8.4           # via -r requirements.in
pyasn1==0.4.8             # via python-jose, rsa
pycparser==2.20           # via cffi
pycryptodomex==3.9.7      # via pyjwkest
pyjwkest==1.4.2           # via drf-oidc-auth
pyjwt==1.7.1              # via djangorestframework-jwt, social-auth-core
pyparsing==2.4.6          # via rdflib
python-dateutil==2.8.1    # via -r requirements.in, icalendar
python-docx==0.8.10       # via -r requirements.in
python-jose==3.1.0        # via django-helusers
python-social-auth==0.3.6  # via -r requirements.in
python3-openid==3.1.0     # via django-allauth, social-auth-core
pytz==2019.3              # via -r requirements.in, django, icalendar
pyyaml==5.3.1             # via -r requirements.in, django-munigeo
rdflib==4.2.2             # via -r requirements.in
requests-cache==0.5.2     # via -r requirements.in, django-munigeo
requests-oauthlib==1.3.0  # via django-allauth, social-auth-core
requests==2.23.0          # via -r requirements.in, django-allauth, django-anymail, django-helusers, django-munigeo, django-orghierarchy, httmock, pyjwkest, requests-cache, requests-oauthlib, social-auth-core
rsa==4.0                  # via python-jose
sentry-sdk==0.14.3        # via -r requirements.in
six==1.14.0               # via -r requirements.in, bleach, cryptography, django-anymail, django-extensions, django-modeltranslation, django-munigeo, django-parler, ecdsa, isodate, langdetect, pyjwkest, python-dateutil, python-jose, social-auth-app-django, social-auth-core
social-auth-app-django==3.1.0  # via -r requirements.in
social-auth-core==3.3.2   # via python-social-auth, social-auth-app-django
sqlparse==0.3.1           # via django
swapper==1.1.2.post1      # via django-orghierarchy
urllib3==1.25.8           # via elasticsearch, requests, sentry-sdk
webencodings==0.5.1       # via bleach

# The following packages are considered to be unsafe in a requirements file:
# setuptools
